refactor(maps): extract useViewport (zoom/pan/transform) from ActiveMapStage
Move the viewport state, container ResizeObserver, fit-scale/effective-scale
transform math, image↔DOM conversion (domToImage), zoom-around-focal-point, and
the non-passive wheel-zoom listener into a useViewport hook. The stage and all
tool hooks now read the transform from one place. No behavior change; verified
by the coordinate-dependent drawing, measurement, and token group-move specs.
